# app/docs/swagger.yaml
definitions:
  aiDataSetModels.AddChunkReq:
    properties:
      content:
        type: string
      dataset_uuid:
        type: string
      document_uuid:
        type: string
      important_keywords:
        items:
          type: string
        type: array
      questions:
        items:
          type: string
        type: array
    type: object
  aiDataSetModels.AgentsCompletionsRequest:
    properties:
      agent_id:
        type: string
      question:
        type: string
      session_id:
        type: string
      stream:
        type: boolean
      sync_dsl:
        type: boolean
      user_id:
        type: string
    type: object
  aiDataSetModels.AskRequest:
    properties:
      dataset_ids:
        items:
          type: string
        type: array
      question:
        type: string
    type: object
...
  aiDataSetModels.ConversationRelatedQuestionsRequest:
    properties:
      question:
        type: string
    type: object
  aiDataSetModels.CreateAssistantRequest:
    properties:
      avatar:
        type: string
      dataset_ids:
        items:
          type: string
        type: array
      description:
        type: string
      icon:
        type: string
      llm:
        $ref: '#/definitions/aiDataSetModels.Llm'
      name:
        type: string
      prompt:
        $ref: '#/definitions/aiDataSetModels.Prompt'
      prompt_config:
        $ref: '#/definitions/aiDataSetModels.PromptConfig'
      top_k:
        type: integer
      top_n:
        type: integer
    type: object
...
  aiDataSetModels.DataSetRequest:
    properties:
      avatar:
        type: string
      chunk_method:
        type: string
      description:
        type: string
      embedding_model:
        type: string
      name:
        type: string
      pagerank:
        type: string
      parser_config:
        $ref: '#/definitions/aiDataSetModels.ParserConfig'
      permission:
        type: string
    required:
    - name
    type: object
  aiDataSetModels.DeleteDocumentRequest:
    properties:
      dataset_id:
        example: "0"
        type: string
      document_ids:
        items:
          type: string
        type: array
    type: object
...
  aiDataSetModels.RemoveAgentSessionsRequest:
    properties:
      agent_id:
        type: string
      ids:
        items:
          type: string
        type: array
    required:
    - agent_id
    type: object
  aiDataSetModels.RemoveChunkReq:
    properties:
      chunk_ids:
        items:
          type: string
        type: array
      dataset_uuid:
        type: string
      document_uuid:
        type: string
    type: object
  aiDataSetModels.RetrievalListReq:
    properties:
      dataset_uuids:
        items:
          type: string
        type: array
      document_uuids:
        items:
          type: string
        type: array
      highlight:
        type: boolean
      keyword:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      question:
        type: string
      rerank_id:
        type: string
      similarity_threshold:
        type: number
      top_k:
        type: integer
      use_kg:
        type: boolean
      vector_similarity_weight:
        type: number
    type: object
  aiDataSetModels.SessionAgentCreate:
    properties:
      agent_id:
        type: string
      data:
        additionalProperties: true
        type: object
    type: object
  aiDataSetModels.UpdateAssistantRequest:
    properties:
      avatar:
        type: string
      chat_id:
        type: string
      dataset_ids:
        items:
          type: string
        type: array
      llm:
        $ref: '#/definitions/aiDataSetModels.Llm'
      name:
        type: string
      prompt:
        $ref: '#/definitions/aiDataSetModels.Prompt'
    type: object
  aiDataSetModels.UpdateChunkReq:
    properties:
      available:
        type: boolean
      chunk_id:
        type: string
      content:
        type: string
      dataset_uuid:
        type: string
      document_uuid:
        type: string
      important_keywords:
        items:
          type: string
        type: array
      questions:
        items:
          type: string
        type: array
    type: object
...
  response.ResponseData:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 相应状态码
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
...
info:
  contact:
    email: zhao_402295440@126.com
    name: danny
  description: baize接口文档
  title: baize
  version: 2.0.x
paths:
  /ai/dataset/agents/list:
    get:
      description: Agen列表
      parameters:
      - in: query
        name: agent_id
        required: true
        type: string
      - in: query
        name: dsl
        type: boolean
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: Agen列表
      tags:
      - 工业智能体/会话管理
  /ai/dataset/assistant/create:
    post:
      description: 创建助理
      parameters:
      - description: 创建助理表请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.CreateAssistantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 创建助理
      tags:
      - 工业智能体/助理管理
  /ai/dataset/assistant/list:
    get:
      description: 读取助理列表
      parameters:
      - in: query
        name: id
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - in: query
        name: name
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 读取助理列表
      tags:
      - 工业智能体/助理管理
  /ai/dataset/assistant/remove/{assistantIds}:
    delete:
      description: 删除助理
      parameters:
      - collectionFormat: csv
        description: assistantIds
        in: path
        items:
          type: string
        name: assistantIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 删除助理
      tags:
      - 工业智能体/助理管理
  /ai/dataset/assistant/update:
    put:
      description: 更新助理
      parameters:
      - description: 更新助理表请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.UpdateAssistantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 更新助理
      tags:
      - 工业智能体/助理管理
  /ai/dataset/chunk/add:
    post:
      description: 添加chunk
      parameters:
      - description: 添加chunk参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.AddChunkReq'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 添加chunk
      tags:
      - 工业智能体/分块管理
  /ai/dataset/chunk/list:
    get:
      description: chunk列表
      parameters:
      - in: query
        name: dataset_uuid
        type: string
      - in: query
        name: document_uuid
        type: string
      - example: ""
        in: query
        name: id
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - example: ""
        in: query
        name: keywords
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: chunk列表
      tags:
      - 工业智能体/分块管理
  /ai/dataset/chunk/remove:
    delete:
      description: 移除chunk
      parameters:
      - description: 移除chunk
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.RemoveChunkReq'
      produces:
      - application/json
      responses:
        "200":
          description: 移除chunk成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 移除chunk
      tags:
      - 工业智能体/分块管理
  /ai/dataset/chunk/retrieval:
    post:
      description: 检索chunk
      parameters:
      - description: 检索chunk请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.RetrievalListReq'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 检索chunk
      tags:
      - 工业智能体/分块管理
  /ai/dataset/chunk/update:
    put:
      description: 更新chunk
      parameters:
      - description: 请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.UpdateChunkReq'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 更新chunk
      tags:
      - 工业智能体/分块管理
...
  /ai/dataset/create:
    post:
      description: 添加数据集
      parameters:
      - description: 设备分组参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.DataSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 添加数据集
      tags:
      - 工业智能体/知识库管理
  /ai/dataset/download/document/{document_id}:
    get:
      description: 下载文档
      parameters:
      - description: document_id
        in: path
        name: document_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 下载文档
      tags:
      - 工业智能体/文档管理
  /ai/dataset/info:
    get:
      description: 读取数据集
      parameters:
      - example: "0"
        in: query
        name: dataset_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 读取数据集
      tags:
      - 工业智能体/知识库管理
  /ai/dataset/list:
    get:
      description: 读取数据集列表
      parameters:
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - example: ""
        in: query
        name: name
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 读取数据集列表
      tags:
      - 工业智能体/知识库管理
  /ai/dataset/list/document:
    get:
      description: 文档列表
      parameters:
      - in: query
        name: dataset_id
        type: integer
      - in: query
        name: document_id
        type: string
      - in: query
        name: document_name
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - in: query
        name: keywords
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 文档列表
      tags:
      - 工业智能体/文档管理
  /ai/dataset/remove:
    delete:
      description: 移除数据集
      parameters:
      - description: dataset_id
        in: path
        name: dataset_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 移除数据集
      tags:
      - 工业智能体/知识库管理
  /ai/dataset/remove/document:
    delete:
      description: 删除文档
      parameters:
      - description: 删除文档请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.DeleteDocumentRequest'
      responses:
        "200":
          description: 删除文档成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 删除文档
      tags:
      - 工业智能体/文档管理
  /ai/dataset/session/agents/completions:
    post:
      description: Agent聊天
      parameters:
      - description: Agent聊天
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.AgentsCompletionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: Agent聊天
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/agents/create:
    post:
      description: 创建agent会话
      parameters:
      - description: 创建agent会话
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.SessionAgentCreate'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 创建agent会话
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/agents/delete:
    delete:
      description: 删除Agent会话
      parameters:
      - description: 删除Agent会话
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.RemoveAgentSessionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 删除Agent会话
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/agents/list:
    get:
      description: Agent会话列表
      parameters:
      - in: query
        name: agent_id
        required: true
        type: string
      - in: query
        name: dsl
        type: boolean
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: Agent会话列表
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/ask:
    post:
      description: 智能问答
      parameters:
      - description: 相关提问
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.AskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 智能问答
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/charts/completions:
    post:
      description: 与聊天助手交谈
      parameters:
      - description: 与聊天助手交谈
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.ChatsCompletionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 与聊天助手交谈
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/conversation/related_questions:
    post:
      description: 相关提问
      parameters:
      - description: 相关提问
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.ConversationRelatedQuestionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 相关提问
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/create:
    post:
      description: 创建助理会话
      parameters:
      - description: 使用 chat 助手创建会话
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.CreateSessionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 创建助理会话
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/list:
    get:
      description: 助理会话列表
      parameters:
      - in: query
        name: chat_id
        type: string
      - in: query
        name: id
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - in: query
        name: name
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 助理会话列表
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/remove:
    delete:
      description: 删除会话
      parameters:
      - description: 删除会话
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.DeleteSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 删除会话
      tags:
      - 工业智能体/会话管理
  /ai/dataset/session/update:
    post:
      description: 更新助理会话
      parameters:
      - description: 更新助理会话
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.UpdateSessionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 更新助理会话
      tags:
      - 工业智能体/会话管理
  /ai/dataset/start/document:
    post:
      description: 解析文档
      parameters:
      - description: 解析文档请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.ParseDocumentApiRequest'
      responses:
        "200":
          description: 解析文档成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 解析文档
      tags:
      - 工业智能体/文档管理
  /ai/dataset/stop/document:
    post:
      description: 停止解析文档
      parameters:
      - description: 停止解析文档参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.ParseDocumentApiRequest'
      responses:
        "200":
          description: 停止解析文档成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 停止解析文档
      tags:
      - 工业智能体/文档管理
  /ai/dataset/update/{dataset_id}:
    put:
      description: 更新数据集
      parameters:
      - description: dataset_id
        in: path
        name: dataset_id
        required: true
        type: string
      - description: 设备分组参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.UpdateDataSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 更新数据集
      tags:
      - 工业智能体/知识库管理
  /ai/dataset/update/document/{document_id}:
    put:
      description: 更新文档
      parameters:
      - description: document_id
        in: path
        name: document_id
        required: true
        type: integer
      - description: 更新文档请求
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/aiDataSetModels.PutDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 更新文档
      tags:
      - 工业智能体/文档管理
  /ai/dataset/upload/document/{datasetId}:
    post:
      consumes:
      - multipart/form-data
      description: 上传文档
      parameters:
      - description: datasetId
        in: path
        name: datasetId
        required: true
        type: integer
      - description: files
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 设置分组成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 上传文档
      tags:
      - 工业智能体/文档管理
...
  /asset/device/template/data/list:
    get:
      description: 获取模板数据列表
      parameters:
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - description: 模板id
        example: "0"
        in: query
        name: template_id
        required: true
        type: string
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 获取模板数据列表
      tags:
      - 资产管理/设备模板管理
  /asset/device/template/data/remove/{ids}:
    delete:
      description: 删除设备数据模板
      parameters:
      - description: ids
        in: path
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 删除设备数据模板
      tags:
      - 资产管理/设备模板管理
...
  /device/monitor/data/list:
    post:
      description: 实时数据
      parameters:
      - description: 实时数据表请求参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/deviceMonitorModel.DevDataReq'
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      summary: 实时数据
      tags:
      - 设备监控/设备监控
...
  /system/dict/data/list:
    get:
      description: 查询字典列表
      parameters:
      - in: query
        name: dictLabel
        type: string
      - in: query
        name: dictType
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysDictDataVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询字典列表
      tags:
      - 字典相关
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"